Added Logger

+++ src/main/java/com/crud/lumencrud/service/ProductService.java
@@ -1,7 +1,10 @@
 package com.crud.lumencrud.service;
 
+import com.crud.lumencrud.controller.ProductController;
 import com.crud.lumencrud.entity.Product;
 import com.crud.lumencrud.repository.ProductRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -9,6 +12,8 @@ import java.util.List;
 
 @Service
 public class ProductService {
+    Logger logger = LoggerFactory.getLogger(ProductService.class);
+
     @Autowired
     private ProductRepository repository;
 
@@ -24,15 +29,26 @@ public class ProductService {
         return repository.findAll();
     }
     public Product getProductbyId(int id){
-        return repository.findById(id).orElse(null);
+        if(repository.existsById(id)){
+            return repository.findById(id).orElse(null);
+        }else{
+            return null;
+        }
+
     }
     public Product getProductbyName(String name){
         return repository.findByName(name);
     }
     public String deleteProduct(int id){
-        repository.deleteById(id);
-        return "product removed ||" +id;
-    }
+        if(repository.existsById(id)){
+            logger.trace("Product Info Deleted");
+            repository.deleteById(id);
+            return "product removed ||" + id;
+        }else{
+            logger.trace("Product Info Not Available to Delete");
+            return "product not available";
+        }
+    }//loggger how to add log
     public  Product updateProduct(Product product){
         Product existingProduct = repository.findById(product.getId()).orElse(null);
         existingProduct.setName(product.getName());
